requirements.txt
cherrypy
dataspyre==0.2.0
